Rearrange ssh logic to bring it together
